celeste/requirements.txt
mss
pillow
